requirements.txt
arrow==1.3.0
attrs==23.2.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
ics==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
load-dotenv==0.1.0
MarkupSafe==2.1.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
TatSu==5.12.1
types-python-dateutil==2.9.0.20240316
watchdog==4.0.1
Werkzeug==3.0.3
